// src/app/generated_services/model/showFeePlanDTO.ts
/**
 * BjjClutch Backend API
 * API for managing students, belts, graduation requirements, notices, frequencies and user authentication with Keycloak
 *
 * Contact: marcosturossi@gmail.com.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ShowFeePlanDTO { 
    id?: string;
    name?: string | null;
    description?: string | null;
    monthDuration?: number;
    price?: number;
    isActive?: boolean;
    createdAt?: string;
}