// src/app/generated_services/model/updateFeePlanDTO.ts
/**
 * BjjClutch Backend API
 * API for managing students, belts, graduation requirements, notices, frequencies and user authentication with Keycloak
 *
 * Contact: marcosturossi@gmail.com.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateFeePlanDTO { 
    name?: string | null;
    description?: string | null;
    monthDuration?: number | null;
    price?: number | null;
    isActive?: boolean | null;
}